Tokens Created
Wasn't sure exactly where to get started so I looked to some of the past
projects for inspiration. The way Morph declared Tokens was the best I
saw. It was so clean and pretty. I took a lot of inspiration from that
because I don't think there is a better way to do it.