Progress on Lexer
Made some more progress on Lexer. Thought I was done but went to test it
and got some weird feedback. If I try to declare an int or string
followed by a character the lexer finds the int or string token but
doesn't acknowledge the space or the character after.